rafts/api/validator/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "raft-validator"
version = "0.1.0"
description = "RAFT ADES Validator"
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn>=0.22.0",
    "lxml>=4.9.2",
    "python-multipart",
    "iau-ades>=0.1.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "httpx>=0.24.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.10.0",
    "pre-commit>=3.5.0",
    "ruff>=0.4.4",
]


[tool.setuptools.packages.find]
include = ["app*"]

[tool.black]
line-length = 88

[tool.ruff]
line-length = 88
